requirements.txt
aiofiles==23.2.1
aiohttp==3.9.1
aiomysql==0.2.0
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.1.0
asgiref==3.7.2
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.1.0
azure-core==1.29.5
azure-identity==1.15.0
blinker==1.7.0
build==1.0.3
CacheControl==0.13.1
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
colorama==0.4.6
crashtest==0.4.1
cryptography==41.0.7
distlib==0.3.8
distro==1.8.0
dnspython==2.4.2
dulwich==0.21.7
exceptiongroup==1.2.0
fastjsonschema==2.19.0
filelock==3.13.1
Flask==3.0.0
frozenlist==1.4.0
greenlet==3.0.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.2
httpx==0.25.2
Hypercorn==0.15.0
hyperframe==6.0.1
idna==3.6
importlib-metadata==7.0.0
installer==0.7.0
isodate==0.6.1
itsdangerous==2.1.2
jaraco.classes==3.3.0
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-path==0.3.2
jsonschema-spec==0.2.4
jsonschema-specifications==2023.7.1
keyring==24.3.0
lazy-object-proxy==1.10.0
MarkupSafe==2.1.3
more-itertools==10.1.0
motor==3.3.2
msal==1.26.0
msal-extensions==1.1.0
msgpack==1.0.7
multidict==6.0.4
numpy==1.26.2
openai==1.3.9
openapi-core==0.18.2
openapi-schema-validator==0.6.2
openapi-spec-validator==0.7.1
packaging==23.2
parse==1.20.0
pathable==0.4.3
pexpect==4.9.0
pkginfo==1.9.6
platformdirs==3.11.0
poetry==1.7.1
poetry-core==1.8.1
poetry-plugin-export==1.6.0
portalocker==2.8.2
prance==23.6.21.0
priority==2.0.0
ptyprocess==0.7.0
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
PyJWT==2.8.0
pymongo==4.6.1
PyMySQL==1.1.0
pyproject_hooks==1.0.0
python-dotenv==1.0.0
#pywin32==306
#pywin32-ctypes==0.2.2
PyYAML==6.0.1
Quart==0.19.4
rapidfuzz==3.5.2
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rpds-py==0.15.2
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
semantic-kernel==0.4.3.dev0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.24
sqlapagination==0.0.1
tenacity==8.0.1
tomlkit==0.12.3
tqdm==4.66.1
trove-classifiers==2023.11.29
typing_extensions==4.9.0
urllib3==2.1.0
virtualenv==20.25.0
Werkzeug==3.0.1
wsproto==1.2.0
yarl==1.9.4
zipp==3.17.0
